Add backlight support
Added the backlight support.  A lot of it came from the example code up on BuyDisplay.com for the display.

Example sketch updated allow me to turn on and to change the pwm brightness for test.

Warning: I scratched my head some yesterday, as nothing worked.  But then found jumpers were set for External on my display.  So un-jumperd J27 and Jumpred J28 and it worked!