WIP - Rotation - Starting to work for 0,1
Added the setRotation and getRotation methods like the other
display drivers.  Not all rotations are working.  And actually
one can not work. but at least we have normal and 90 degree
partially working.

Note: to get portrait mode to work, like the RA8875 you have to
swap X and Y values in many places.  This has not been fully
completed yet.  I did do rectangle functions, line, and tested
some text output (Note this is only text that we output and not
the built in text of the display

So still WIP. But Progres.

Test sketch has been updated to try to cycle through the
different rotations.