Quick and dirty writeRect implementation plus example sketch
This is a port of my RA8875 writeRect test sketch plus

a version of writeRect, which uses the bteMpuWriteWithROPData code to draw the image.

In most rotations (other than 2) avoided having to copy the data, instead the different bte operations were adjusted to write out either one row or one column.

Only in Rotation 2, did I need to malloc a buffer to reverse the bytes in order to output.

Test sketch was added that displays 4 images with 5 second delay between
and then it switches to next rotation.

You can switch from 5 seconds to your own pace by typing in a:
```
s
```
command at the prompt and then you can step one page at a time.

Unlike other versions of similar test sketch this one prints two numbers
per page.  The time it took to call the function which in some case only
is up to when the DMA operation started.  And then the second number is
the time after it displayed the delta time on the screen.  As the code
is setup to wait until any pending DMA operations completes before it
outputs to screen, this shows the actual time to complete.

May need to watch rotation 2 to make sure that we are not overwriteing
memory before the DMA gets a chance to output... May need to allocate 2
buffers and alternate.