Experiment - with allowing you to see if pre-rotating an image speeds up...
The writeRect code has to either manipluate the bits going out and/or it has to only output one row or column at a time.

So hacked up a version of the code to malloc a new bitmap where I reorder the bytes such that you can do one output operation.

Test sketch does  my larger T4.1 cdard and draws it again.

Wondering about some of the bits not fully drawing toward bottom of sketch?