Make multi-resolutions version work on 1024x600 display
The version that was setup to work on 1280x400 display did not display correctly on the 1024x600 displays.  Most obvious with the draw text 4 sketch which first lines of text were off the screen.

Fixed (maybe) 3 things:
a) Status text - was adjusting height twice, not sure if was issue or not
b) calculations for Page starting address was off by a page.  Page(1) should be 0...
c) There are timing settings different between the two displays.  I made two copies in the setting file and the init code decided on width if it should use the 1280 displays setting or default to what we had...

Note: I left in one difference where before XPCLK_INV was 1 but with this branch it is 0...  I left it with new value, but should watch out for it.  Note: This one value is used in 4 places.

So far it appears to work on my my 1024x600 display