Update RA8876_t3.cpp
bteMpuWriteColorExpansionData() and bteMpuWriteColorExpansionWithChromaKeyData() do not currently correctly handle the case where the width is not an exact multiple of 8. This change rounds up the width to the nearest exact byte boundary to ensure that the gpu gets enough data to process the command. If the width is an exact multiple of 8 then this code will behave exactly as before.